# backend/api/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"github.com/go-faster/errors"
)

// CreateEditAccepted is response for CreateEdit operation.
type CreateEditAccepted struct{}

func (*CreateEditAccepted) createEditRes() {}

// CreateEditBadRequest is response for CreateEdit operation.
type CreateEditBadRequest struct{}

func (*CreateEditBadRequest) createEditRes() {}

// Ref: #/components/schemas/Edit
type Edit struct {
	ID        int64        `json:"id"`
	Name      string       `json:"name"`
	Author    string       `json:"author"`
	Prompt    string       `json:"prompt"`
	Start     GeoJsonPoint `json:"start"`
	End       GeoJsonPoint `json:"end"`
	CreatedAt string       `json:"createdAt"`
	ImagePath string       `json:"imagePath"`
}

// GetID returns the value of ID.
func (s *Edit) GetID() int64 {
	return s.ID
}

// GetName returns the value of Name.
func (s *Edit) GetName() string {
	return s.Name
}

// GetAuthor returns the value of Author.
func (s *Edit) GetAuthor() string {
	return s.Author
}

// GetPrompt returns the value of Prompt.
func (s *Edit) GetPrompt() string {
	return s.Prompt
}

// GetStart returns the value of Start.
func (s *Edit) GetStart() GeoJsonPoint {
	return s.Start
}

// GetEnd returns the value of End.
func (s *Edit) GetEnd() GeoJsonPoint {
	return s.End
}

// GetCreatedAt returns the value of CreatedAt.
func (s *Edit) GetCreatedAt() string {
	return s.CreatedAt
}

// GetImagePath returns the value of ImagePath.
func (s *Edit) GetImagePath() string {
	return s.ImagePath
}

// SetID sets the value of ID.
func (s *Edit) SetID(val int64) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *Edit) SetName(val string) {
	s.Name = val
}

// SetAuthor sets the value of Author.
func (s *Edit) SetAuthor(val string) {
	s.Author = val
}

// SetPrompt sets the value of Prompt.
func (s *Edit) SetPrompt(val string) {
	s.Prompt = val
}

// SetStart sets the value of Start.
func (s *Edit) SetStart(val GeoJsonPoint) {
	s.Start = val
}

// SetEnd sets the value of End.
func (s *Edit) SetEnd(val GeoJsonPoint) {
	s.End = val
}

// SetCreatedAt sets the value of CreatedAt.
func (s *Edit) SetCreatedAt(val string) {
	s.CreatedAt = val
}

// SetImagePath sets the value of ImagePath.
func (s *Edit) SetImagePath(val string) {
	s.ImagePath = val
}

// Ref: #/components/schemas/EditRequest
type EditRequest struct {
	// Human-readable name for this edit.
	Name string `json:"name"`
	// Name or identifier of the person submitting the edit.
	Author string `json:"author"`
	// Free-text description of the desired edit.
	Prompt string       `json:"prompt"`
	Start  GeoJsonPoint `json:"start"`
	End    GeoJsonPoint `json:"end"`
}

// GetName returns the value of Name.
func (s *EditRequest) GetName() string {
	return s.Name
}

// GetAuthor returns the value of Author.
func (s *EditRequest) GetAuthor() string {
	return s.Author
}

// GetPrompt returns the value of Prompt.
func (s *EditRequest) GetPrompt() string {
	return s.Prompt
}

// GetStart returns the value of Start.
func (s *EditRequest) GetStart() GeoJsonPoint {
	return s.Start
}

// GetEnd returns the value of End.
func (s *EditRequest) GetEnd() GeoJsonPoint {
	return s.End
}

// SetName sets the value of Name.
func (s *EditRequest) SetName(val string) {
	s.Name = val
}

// SetAuthor sets the value of Author.
func (s *EditRequest) SetAuthor(val string) {
	s.Author = val
}

// SetPrompt sets the value of Prompt.
func (s *EditRequest) SetPrompt(val string) {
	s.Prompt = val
}

// SetStart sets the value of Start.
func (s *EditRequest) SetStart(val GeoJsonPoint) {
	s.Start = val
}

// SetEnd sets the value of End.
func (s *EditRequest) SetEnd(val GeoJsonPoint) {
	s.End = val
}

// Ref: #/components/schemas/GeoJsonPoint
type GeoJsonPoint struct {
	Type GeoJsonPointType `json:"type"`
	// [longitude, latitude] in WGS84.
	Coordinates []float64 `json:"coordinates"`
}

// GetType returns the value of Type.
func (s *GeoJsonPoint) GetType() GeoJsonPointType {
	return s.Type
}

// GetCoordinates returns the value of Coordinates.
func (s *GeoJsonPoint) GetCoordinates() []float64 {
	return s.Coordinates
}

// SetType sets the value of Type.
func (s *GeoJsonPoint) SetType(val GeoJsonPointType) {
	s.Type = val
}

// SetCoordinates sets the value of Coordinates.
func (s *GeoJsonPoint) SetCoordinates(val []float64) {
	s.Coordinates = val
}

type GeoJsonPointType string

const (
	GeoJsonPointTypePoint GeoJsonPointType = "Point"
)

// AllValues returns all GeoJsonPointType values.
func (GeoJsonPointType) AllValues() []GeoJsonPointType {
	return []GeoJsonPointType{
		GeoJsonPointTypePoint,
	}
}

// MarshalText implements encoding.TextMarshaler.
func (s GeoJsonPointType) MarshalText() ([]byte, error) {
	switch s {
	case GeoJsonPointTypePoint:
		return []byte(s), nil
	default:
		return nil, errors.Errorf("invalid value: %q", s)
	}
}

// UnmarshalText implements encoding.TextUnmarshaler.
func (s *GeoJsonPointType) UnmarshalText(data []byte) error {
	switch GeoJsonPointType(data) {
	case GeoJsonPointTypePoint:
		*s = GeoJsonPointTypePoint
		return nil
	default:
		return errors.Errorf("invalid value: %q", data)
	}
}